refactor(auth): relocate auth-profiles from src/agents/ to src/auth/ (#419)
Move auth-profile modules out of the agent-specific scope into a
dedicated top-level src/auth/ directory, reflecting their middleware-wide
role in the unified auth architecture.

Moved files:
- src/agents/auth-profiles/ → src/auth/ (flattened)
- src/agents/auth-profiles.ts → src/auth/index.ts (barrel)
- src/agents/provider-auth.ts → src/auth/provider-auth.ts
- src/agents/api-key-rotation.ts → src/auth/api-key-rotation.ts
- src/agents/auth-health.ts → src/auth/auth-health.ts

All consumer imports, vi.mock() paths, and internal relative imports
updated. On-disk storage path unchanged. No behavioral changes.
