--- src/agents/auth-profiles.ts
@@ -1,20 +1,0 @@
-export { CLAUDE_CLI_PROFILE_ID, CODEX_CLI_PROFILE_ID } from "./auth-profiles/constants.js";
-export { resolveAuthProfileDisplayLabel } from "./auth-profiles/display.js";
-export { formatAuthDoctorHint } from "./auth-profiles/doctor.js";
-export { resolveApiKeyForProfile } from "./auth-profiles/oauth.js";
-export { resolveAuthStorePathForDisplay } from "./auth-profiles/paths.js";
-export {
-  listProfilesForProvider,
-  upsertAuthProfile,
-  upsertAuthProfileWithLock,
-} from "./auth-profiles/profiles.js";
-export {
-  ensureAuthProfileStore,
-  loadAuthProfileStore,
-  saveAuthProfileStore,
-} from "./auth-profiles/store.js";
-export type {
-  ApiKeyCredential,
-  AuthProfileCredential,
-  AuthProfileStore,
-} from "./auth-profiles/types.js";
