+++ src/auth/index.ts
@@ -1,0 +1,12 @@
+export { CLAUDE_CLI_PROFILE_ID, CODEX_CLI_PROFILE_ID } from "./constants.js";
+export { resolveAuthProfileDisplayLabel } from "./display.js";
+export { formatAuthDoctorHint } from "./doctor.js";
+export { resolveApiKeyForProfile } from "./oauth.js";
+export { resolveAuthStorePathForDisplay } from "./paths.js";
+export {
+  listProfilesForProvider,
+  upsertAuthProfile,
+  upsertAuthProfileWithLock,
+} from "./profiles.js";
+export { ensureAuthProfileStore, loadAuthProfileStore, saveAuthProfileStore } from "./store.js";
+export type { ApiKeyCredential, AuthProfileCredential, AuthProfileStore } from "./types.js";
