+++ src/auth/paths.ts
@@ -1,8 +1,8 @@
 import fs from "node:fs";
 import path from "node:path";
-import { saveJsonFile } from "../../infra/json-file.js";
-import { resolveUserPath } from "../../utils.js";
-import { resolveRemoteClawAgentDir } from "../agent-paths.js";
+import { resolveRemoteClawAgentDir } from "../agents/agent-paths.js";
+import { saveJsonFile } from "../infra/json-file.js";
+import { resolveUserPath } from "../utils.js";
 import { AUTH_PROFILE_FILENAME, AUTH_STORE_VERSION, LEGACY_AUTH_FILENAME } from "./constants.js";
 import type { AuthProfileStore } from "./types.js";
