+++ src/commands/agent.ts
@@ -3,13 +3,13 @@ import {
   resolveSessionAgentId,
   resolveAgentWorkspaceDir,
 } from "../agents/agent-scope.js";
-import { ensureAuthProfileStore } from "../agents/auth-profiles.js";
-import { clearSessionAuthProfileOverride } from "../agents/auth-profiles/session-override.js";
 import { resolveChannelMessageToolHints } from "../agents/channel-tools.js";
 import { getCliSessionId } from "../agents/cli-session.js";
 // Model management defaults gutted in RemoteClaw — CLI runtimes own model selection.
 import { normalizeModelRef } from "../agents/provider-utils.js";
 import { ensureAgentWorkspace } from "../agents/workspace.js";
+import { ensureAuthProfileStore } from "../auth/index.js";
+import { clearSessionAuthProfileOverride } from "../auth/session-override.js";
 import { normalizeVerboseLevel, type VerboseLevel } from "../auto-reply/thinking.js";
 import { formatCliCommand } from "../cli/command-format.js";
 import { type CliDeps, createDefaultDeps } from "../cli/deps.js";
