+++ src/commands/channels.adds-non-default-telegram-account.test.ts
@@ -7,8 +7,8 @@ const authMocks = vi.hoisted(() => ({
   loadAuthProfileStore: vi.fn(),
 }));
 
-vi.mock("../agents/auth-profiles.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../agents/auth-profiles.js")>();
+vi.mock("../auth/index.js", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("../auth/index.js")>();
   return {
     ...actual,
     loadAuthProfileStore: authMocks.loadAuthProfileStore,
