+++ src/commands/doctor.e2e-harness.ts
@@ -226,8 +226,8 @@ vi.mock("../infra/update-runner.js", () => ({
   runGatewayUpdate,
 }));
 
-vi.mock("../agents/auth-profiles.js", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("../agents/auth-profiles.js")>();
+vi.mock("../auth/index.js", async (importOriginal) => {
+  const actual = await importOriginal<typeof import("../auth/index.js")>();
   return {
     ...actual,
     ensureAuthProfileStore,
